refactor: modularize authentication components
